+++ README.md
@@ -72,8 +72,7 @@ node main.js --noedit --diff
 
 ## TODO
 
-[ ] 完善域名列表  
-[ ] 将写入 `HOSTS` 文件的代码和获取 IP 的代码分离  
+[ ] 完善域名列表   
 [ ] 使其可以用于所有这样的网站  
 [ ] 优化代码,增加可读性  
 [ ] 简化步骤，试图让用户一键配置，跑完代码
